Interface changes

+++ protobuf_gen/wrap.py
@@ -7,7 +7,7 @@ import sys
 from collections import __init__
 from importlib import import_module
 from tempfile import mkdtemp
-from typing import List, Dict
+from typing import List, Dict, Optional, Tuple
 
 from google.protobuf import descriptor_pool, descriptor_database
 
@@ -20,18 +20,23 @@ def _clear_pb_descriptor_database():
     """
     !!! This is a dirty hack, but ensures that tests can both run the compiler and then re-import the files again.
     """
-    dp = descriptor_pool._DEFAULT
+    print('Use c descriptors', descriptor_pool._USE_C_DESCRIPTORS)
 
-    dp._internal_db = descriptor_database.DescriptorDatabase()
-    dp._descriptor_db = None
-    dp._descriptors = {}
-    dp._enum_descriptors = {}
-    dp._service_descriptors = {}
-    dp._file_descriptors = {}
-    dp._toplevel_extensions = {}
-    dp._file_desc_by_toplevel_extension = {}
-    dp._extensions_by_name = collections.defaultdict(dict)
-    dp._extensions_by_number = collections.defaultdict(dict)
+    if not descriptor_pool._USE_C_DESCRIPTORS:
+        dp = descriptor_pool._DEFAULT
+
+        dp._internal_db = descriptor_database.DescriptorDatabase()
+        dp._descriptor_db = None
+        dp._descriptors = {}
+        dp._enum_descriptors = {}
+        dp._service_descriptors = {}
+        dp._file_descriptors = {}
+        dp._toplevel_extensions = {}
+        dp._file_desc_by_toplevel_extension = {}
+        dp._extensions_by_name = collections.defaultdict(dict)
+        dp._extensions_by_number = collections.defaultdict(dict)
+    else:
+        pass
 
 
 def _random_pkg_name():
@@ -42,26 +47,35 @@ def _load_protoc_mods(
     includes: List[str],
     input_proto: List[str],
     output_files: Dict[str, InputModule],
+    workdir: str = None,
+    pkg: str = None,
+    existing=False
 ):
-    workdir = mkdtemp()
-    pkg = _random_pkg_name()
+    if workdir is None:
+        workdir = mkdtemp()
+    if pkg is None:
+        pkg = _random_pkg_name()
 
-    protoc_output_dir = os.path.join(workdir, pkg)
+    protoc_output_dir = os.path.join(workdir, pkg.replace('.', '/'))
 
     os.makedirs(protoc_output_dir, exist_ok=True)
 
     try:
-        sys.path = sys.path + [workdir]
+        sys.path = sys.path + [os.path.join(workdir)]
 
-        _build_pb_with_prefix(
-            pkg,
-            protoc_output_dir,
-            includes,
-            input_proto
-        )
+        if not existing:
+
+            _build_pb_with_prefix(
+                pkg,
+                protoc_output_dir,
+                includes,
+                input_proto
+            )
 
         output_mods: Dict[str, OutputModule] = {}
         clear_mods = []
+
+        print('Workdir', workdir)
 
         for k, of in output_files.items():
             to_import = pkg + '.' + of.mod
@@ -72,12 +86,15 @@ def _load_protoc_mods(
 
             output_mods[k] = of.to_output(m.DESCRIPTOR)
 
-        _clear_pb_descriptor_database()
+        if not existing:
+            _clear_pb_descriptor_database()
 
         return output_mods
     finally:
         sys.path = [x for x in sys.path if x != workdir]
-        shutil.rmtree(workdir)
+
+        if not existing:
+            shutil.rmtree(workdir)
 
 
 def wrap(
@@ -86,7 +103,8 @@ def wrap(
     root_autogen='autogen',
     includes: List[str] = None,
     input_proto: List[str] = None,
-    output_files: Dict[str, InputModule] = None
+    output_files: Dict[str, InputModule] = None,
+    existing: Optional[Tuple[str, str]] = None
 ):
     build(
         BuildProps(
@@ -95,7 +113,8 @@ def wrap(
             _load_protoc_mods(
                 includes,
                 input_proto,
-                output_files
+                output_files,
+                *(existing + (True,) if existing else (None, None, False))
             )
         ),
         outdir=output_dir_wrappers
